support non distributed build on mac

+++ src/feature/Derivatives.cpp
@@ -75,7 +75,7 @@ std::vector<T> Derivatives<T>::computeDerivative(
       size_t curIdx = i * numfeat + j;
       for (size_t d = 1; d <= windowlen; ++d) {
         output[curIdx] += d *
-            (input[curIdx + std::min((numframes - i - 1), d) * numfeat] -
+            (input[curIdx + std::min((numframes - i - 1), (uint64_t) d) * numfeat] -
              input[curIdx - std::min(i, d) * numfeat]);
       }
       output[curIdx] /= denominator;
